renamed index to login and register
