+++ src/main.tsx
@@ -7,9 +7,9 @@ import { Provider } from 'react-redux';
 import { RootState, store } from './services/redux/store';
 import { useAppDispatch, useAppSelector } from './services/typeHooks';
 import { ROUTE_HOME, ROUTE_LOGIN, ROUTE_REGISTER } from './utils/constants';
-import LoginPage from './pages/login';
+import LoginPage from './pages/login/login';
 import NotFound404 from './pages/notfound404/notfound404';
-import RegisterPage from './pages/register';
+import RegisterPage from './pages/register/register';
 import Loader from './components/loader';
 import { logoutUser } from './services/redux/slices/auth/auth';
